feat(projects): add layout shell
